requirements.txt
# This file is used to specify dependencies for the project.

# dependencies for running the application
paho-mqtt==1.6.1
Flask==3.0.0

# dependencies for code quality and linting in CI/CD pipeline
isort==5.13.2
pylint==3.0.3
black==23.12.1

pytest==7.4.4
pytest-cov==4.1.0


